refactor: export Pollable in experimental/sys, keep fsapi internal
Address review feedback on PR #2462:

- Export Pflag, POLLIN, POLLOUT, and Pollable interface in
  experimental/sys so users can implement custom Poll() on their
  io.Reader passed to WithStdin().
- Move experimental/fsapi back to internal/fsapi since the extended
  File interface is only used internally. internal/fsapi re-exports
  Pflag types from experimental/sys via type aliases.
- Replace duplicate pollable interface definitions in internal/sys
  and internal/sysfs with the exported experimental/sys.Pollable.
- Fix poll check ordering in StdinFile.Poll: check Pollable on the
  reader before the POLLIN flag check, so custom implementations
  can handle non-POLLIN flags.

No backwards-breaking changes to pre-existing public API.
experimental/sys.File interface is unchanged.
