+++ internal/fsapi/poll.go
@@ -1,5 +1,7 @@
 package fsapi
 
+import experimentalsys "github.com/tetratelabs/wazero/experimental/sys"
+
 // Pflag are bit flags used for File.Poll. Values, including zero, should not
 // be interpreted numerically. Instead, use by constants prefixed with 'POLL'.
 //
@@ -7,14 +9,14 @@ package fsapi
 //
 //   - This is like `pollfd.events` flags for `poll` in POSIX. See
 //     https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/poll.h.html
-type Pflag uint32
+type Pflag = experimentalsys.Pflag
 
 // Only define bitflags we support and are needed by `poll_oneoff` in wasip1
 // See https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md#eventrwflags
 const (
 	// POLLIN is a read event.
-	POLLIN Pflag = 1 << iota
+	POLLIN = experimentalsys.POLLIN
 
 	// POLLOUT is a write event.
-	POLLOUT
+	POLLOUT = experimentalsys.POLLOUT
 )
