+++ internal/sys/stdio.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	experimentalsys "github.com/tetratelabs/wazero/experimental/sys"
-	"github.com/tetratelabs/wazero/experimental/fsapi"
+	"github.com/tetratelabs/wazero/internal/fsapi"
 	"github.com/tetratelabs/wazero/internal/sysfs"
 	"github.com/tetratelabs/wazero/sys"
 )
@@ -24,19 +24,14 @@ func (f *StdinFile) Read(buf []byte) (int, experimentalsys.Errno) {
 	return n, experimentalsys.UnwrapOSError(err)
 }
 
-// pollable has just the Poll function.
-type pollable interface {
-	Poll(fsapi.Pflag, int32) (ready bool, errno experimentalsys.Errno)
-}
-
 // Poll implements the same method as documented on fsapi.File
 func (f *StdinFile) Poll(flag fsapi.Pflag, timeoutMillis int32) (ready bool, errno experimentalsys.Errno) {
-	if flag != fsapi.POLLIN {
-		return false, experimentalsys.ENOTSUP
+	if p, ok := f.Reader.(experimentalsys.Pollable); ok {
+		return p.Poll(flag, timeoutMillis)
 	}
 
-	if p, isPollable := f.Reader.(pollable); isPollable {
-		return p.Poll(flag, timeoutMillis)
+	if flag != fsapi.POLLIN {
+		return false, experimentalsys.ENOTSUP
 	}
 
 	return true, 0 // always ready to read nothing
