+++ internal/sysfs/sock_unsupported.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/tetratelabs/wazero/experimental/sys"
 	experimentalsys "github.com/tetratelabs/wazero/experimental/sys"
-	"github.com/tetratelabs/wazero/experimental/fsapi"
+	"github.com/tetratelabs/wazero/internal/fsapi"
 	socketapi "github.com/tetratelabs/wazero/internal/sock"
 )
