+++ internal/sysfs/sock_windows.go
@@ -10,7 +10,7 @@ import (
 	"golang.org/x/sys/windows"
 
 	"github.com/tetratelabs/wazero/experimental/sys"
-	"github.com/tetratelabs/wazero/experimental/fsapi"
+	"github.com/tetratelabs/wazero/internal/fsapi"
 	socketapi "github.com/tetratelabs/wazero/internal/sock"
 )
